refactor: prayerUpdate Dir 구조 변경 #25
- prayers로 통일
