+++ app/prayers/[id]/_components/update/PrayerUpdateBody.tsx
@@ -2,26 +2,26 @@ import React, { useState, useRef, useMemo, useEffect } from "react";
 import { View, StyleSheet, Platform, Dimensions } from "react-native";
 import { Divider, Text } from "react-native-paper";
 import { RFValue } from "react-native-responsive-fontsize";
-import { color } from "../../../common/styles/color";
-import { SelectedMember } from "../../prayers/types/SelectedMember";
+import { color } from "../../../../../src/common/styles/color";
+import { SelectedMember } from "../../../../../src/domain/prayers/types/SelectedMember";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
-import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
-import { useRoomMembersQuery } from "../../rooms/hooks/queries/roomQueries";
-import { RoomMember } from "../../rooms/types/roomMember";
-import { PrayerUpdateItem } from "../types/prayerUpdateItem";
-import { usePrayerUpdateStore } from "../stores/usePrayerUpdateStore";
-import { useSelectedPrayerTitleStore } from "../../prayers/stores/useSelectedPrayerTitleStore";
-import { usePrayerContentsQuery } from "../../prayers/hooks/queries/usePrayerQueries";
+import { useSelectedRoomStore } from "../../../../../src/domain/rooms/stores/useSelectedRoomStore";
+import { useRoomMembersQuery } from "../../../../../src/domain/rooms/hooks/queries/roomQueries";
+import { RoomMember } from "../../../../../src/domain/rooms/types/roomMember";
+import { PrayerUpdateItem } from "../../../../../src/domain/prayers/types/prayerUpdateItem";
+import { usePrayerUpdateStore } from "../../../../../src/domain/prayers/stores/usePrayerUpdateStore";
+import { useSelectedPrayerTitleStore } from "../../../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
+import { usePrayerContentsQuery } from "../../../../../src/domain/prayers/hooks/queries/usePrayerQueries";
 
-import PrayerTitleInput from "../../../../app/prayers/_components/PrayerTitleInput";
-import PrayerMemberSelector from "../../../../app/prayers/_components/PrayerMemberSelector";
-import PrayerContentInput from "../../../../app/prayers/_components/PrayerContentInput";
-import PrayerAddButton from "../../../../app/prayers/_components/PrayerAddButton";
-import PrayerCarousel from "../../../../app/prayers/_components/PrayerCarousel";
-import PrayerDeleteDialog from "../../../../app/prayers/_components/dialog/PrayerDeleteDialog";
-import PrayerCustomNameDialog from "../../../../app/prayers/_components/dialog/PrayerCustomNameDialog";
-import PrayerMemberSelectionModal from "../../../../app/prayers/_components/modal/PrayerMemberSelectionModal";
-import PrayerEditDialog from "../../../../app/prayers/_components/dialog/PrayerEditDialog";
+import PrayerTitleInput from "../../../_components/PrayerTitleInput";
+import PrayerMemberSelector from "../../../_components/PrayerMemberSelector";
+import PrayerContentInput from "../../../_components/PrayerContentInput";
+import PrayerAddButton from "../../../_components/PrayerAddButton";
+import PrayerCarousel from "../../../_components/PrayerCarousel";
+import PrayerDeleteDialog from "../../../_components/dialog/PrayerDeleteDialog";
+import PrayerCustomNameDialog from "../../../_components/dialog/PrayerCustomNameDialog";
+import PrayerMemberSelectionModal from "../../../_components/modal/PrayerMemberSelectionModal";
+import PrayerEditDialog from "../../../_components/dialog/PrayerEditDialog";
 
 const { width } = Dimensions.get("window");
 const CARD_SPACING = RFValue(20); // 카드 사이 간격
