+++ src/domain/prayers/types/params/updatePrayerParams.ts
@@ -1,0 +1,8 @@
+import { PrayerUpdateItem } from "./../prayerUpdateItem";
+
+export interface UpdatePrayerParams {
+  roomId: number | null;
+  prayerTitleId: number | null;
+  title: string;
+  prayerList: PrayerUpdateItem[];
+}
