+++ src/domain/prayers/types/prayerUpdateItem.ts
@@ -1,0 +1,3 @@
+import type { PrayerCreationItem } from "./PrayerCreationItem";
+
+export interface PrayerUpdateItem extends PrayerCreationItem {}
