+++ src/domain/prayers/types/prayerUpdateStore.ts
@@ -1,10 +1,17 @@
 import { PrayerUpdateItem } from "./prayerUpdateItem";
 
-export interface PrayerUpdateList {
+export interface PrayerUpdateState {
   prayerList: PrayerUpdateItem[];
+}
+
+export interface PrayerUpdateActions {
   add: (prayer: PrayerUpdateItem) => void;
   delete: (prayer: PrayerUpdateItem) => void;
   update: (prayer: PrayerUpdateItem) => void;
   set: (prayers: PrayerUpdateItem[]) => void;
   clear: () => void;
 }
+
+export interface PrayerUpdateStore
+  extends PrayerUpdateState,
+    PrayerUpdateActions {}
