+++ src/domain/prayers/types/request/updatePrayerRequest.ts
@@ -1,6 +1,6 @@
 import { PrayerUpdateItem } from "../prayerUpdateItem";
 
-export interface PrayerUpdateRequest {
+export interface UpdatePrayerRequest {
   prayers: {
     title: string;
     contents: PrayerUpdateItem[];
